feat(client): add commitButton component and use it from commit and commitsPage components
