+++ packages/amplication-client/src/VersionControl/CommitsPage.tsx
@@ -8,6 +8,7 @@ import { AppRouteProps } from "../routes/routesUtil";
 import CommitList from "./CommitList";
 import "./CommitsPage.scss";
 import BuildPage from "./BuildPage";
+import CommitButton from "./CommitButton";
 
 type Props = AppRouteProps & {
   match: match<{
@@ -86,6 +87,7 @@ const CommitsPage: React.FC<Props> = ({ moduleClass, innerRoutes }) => {
             message="There are no commits to show. "
             image={EnumImages.CommitEmptyState}
           />
+          <CommitButton />
         </>
       )}
     </PageContent>
