+++ packages/amplication-client/src/VersionControl/hooks/commitQueries.ts
@@ -121,3 +121,26 @@ export const GET_COMMIT = gql`
     }
   }
 `;
+
+export const COMMIT_CHANGES = gql`
+  mutation commit(
+    $message: String!
+    $projectId: String!
+    $bypassLimitations: Boolean
+  ) {
+    commit(
+      data: {
+        message: $message
+        bypassLimitations: $bypassLimitations
+        project: { connect: { id: $projectId } }
+      }
+    ) {
+      id
+      builds {
+        id
+        resourceId
+        status
+      }
+    }
+  }
+`;
